requirements.txt
fastapi
uvicorn
pillow
reportlab